Move all components into one directory
